More bitness for food
There probably could be more than 255 food on the map, and there is
no reason why it can't all be brought to one cell.

+++ src/geography.rs
@@ -46,4 +46,4 @@ impl Markers {
 }
 
 #[derive(Clone, Debug)]
-pub struct Food(pub u8);
+pub struct Food(pub u16);
